Fixes Become and Unbecome to work as Akka.
Become(receive, discardOld: true) means that the current receive delegate is replaced by the specified delegate
Become(receive, discardOld: false) means that the current receive delegate is persisted before the the specified delegate is put in place. To revert to previously persisted delegates Unbecome() may be called.
discardOld is an optional parameter which defaults to true.