Add Router.RemoveRoutee & implicit casting ActorRef -> ActorRefRoutee
Simplify AddRoutee, that has the same logic copy-pasted.

Breaking change: Router.Send should not be public. This can never have been meant to be public and I can't see why anyone would actually use it from the outside. The public send-to-routees is Route(). The impact of this breaking change should be slim to nothing.