Use preReleaseVersion for Akka.Cluster
When building:
    build.cmd Nuget
the package name for cluster will be like this:
   Akka.Cluster.0.6.5-pre1410160612

The numbers after pre are the UTC date and time on the format "yyMMddHHmm" to get an increasing version number of the prerelease package (as long as you build the next version the minute after).

This commit also makes sure that when building pre-release versions of all packages using
    build.cmd Nuget nugetprerelease=dev
the name of Akka.Cluster package use the specified prefix. In this example the name will be:
   Akka.Cluster.0.6.5-dev1410160612