Move SerilogLogMessageFormatter to correct namespace.
This should have been made in #292 but we missed that.
It should be backwards compatible, sort of . As long as no
   using Akka.Logger.Serilog;
statement exists in files where SerilogLogMessageFormatter  is used everything should compile (with a warning that SerilogLogMessageFormatter  in the old namespace is obsolete).

If a both namespacese are used:
   using Akka.Logger.Serilog;
   using Akka.Serilog.Event.Serilog;
The user will get a compile error. The resolution is to remove the second using statement.