Replace the unclear MakeDecider with clearer versions
Akka JVM has SupervisorStrategy.makeDecider functions which have been ported to our code.

In our case only one overload has been ported:
    SupervisorStrategy.MakeDecider(exceptions)

When you see the line above in the code it is very unclear, just by looking at the name, what it does.
It returns Restart if the reason to why an actor failed is one of the specified exceptions, and Escalate otherwise.

A more suitable name would probably have been something like
    SupervisorStrategy.MakeRestartDeciderFor(exceptions).

However in our code base it's only used twice, and the usage looks like this:
    SupervisorStrategy.MakeDecider(typeof(Exception))

Which means, always Restart. This commit replaces those calls with the much clearer
    reason => Directive.Restart

If we want to introduce helper functions for making deciders we should not port Akka JVMs way, but introduce our own