Log ActorInitializationException instead of underlying exception
We currently log the underlying exception of an ActorInitializationException since I screwed up porting this from AKKA JVM a week back. :)
We should only use the message from the underlying exception, and still log the entire ActorInitializationException.

This is important since it's otherwise impossible to test that ActorInitializationException is thrown.