Removed a bunch of redundant type checks
Most of these are in the form of the following:

if (message is ClusterUserAction.JoinTo)
{
    var jt = message as ClusterUserAction.JoinTo;
    ...
}

Here message is being checked twice for being of type
ClusterUserAction.JoinTo

So instead I replaced the 'as' cast to a direct cast.

if (message is ClusterUserAction.JoinTo)
{
    var jt = (ClusterUserAction.JoinTo)message;
    ...
}

I would have done the more familiar 'as' cast/check for
null convention but most of these are part of a long
if/else chain that doesn't make sense to break apart.