Introduce IScheduler and use it everywhere instead of Scheduler.
The reason for this commit is to prepare for a scheduler that can be
used in tests without having to run the tests in real time.

ICancelable & Cancelable is basically a wrapper around
CancellationTokenSource. CTS is replaced to be able to
schedule CancelAfter calls on the scheduler, instead of what CTS
normally uses.
If the scheduler is using the normal system time then what CTS
normally uses will be used.

See issue #468 for more information.