Create behavior stack with capacity 1, i.e. new Stack<Receive>(1)
The behavior stack will _always_ contain at least 1 item, as all
actors, the first thing they do, is to call Become, which pushes an item to
the stack.

On first call to Become, as the stack has 0 capacity, a new array with
size 4 is created.
As most actors do not use Become-stacked to push more items on the stack,
this means we're wasting bytes.

If we create the behavior stack with initial size 1, we avoid the
creation of a new array on the first Become-call and for most cases the
underlying array isn't larger than it has to be.