Fix Nullable Supervisor Strategy constructors
In chaining to other constructors, the constructors for the `AllForOne`
and `OneForOne` strategies pull timeouts from `TimeSpan`s using the
`Milliseconds` property. This is not the correct behavior for `TimeSpan`
values greater than or equal to 1 second as it provides only the
milliseconds portion of the timespan. Instead, use the `TotalMilliseconds`
property and cast it to an `int`.