Multinode test runner alterations
Refactored NodeTestRunner to allow MultinodeTestRunner to call it in
two different manners.

`build multinodetests` works in the same manner as in the past. Starting
a seperate process for each node that needs to be created.

Using `build multinodetests spec=<name of spec>` will limit the runner to
starting just the named test. In addition, the nodes will be spun up
inside isolated AppDomains rather than as seperate processes. The final
difference between single test and multi test mode, is that the runner
will call `Debugger.Launch()` to attach VS to the test.

By automatically launching the debugger, I've been able to remove
the `Thread.Sleep()`s that were needed previously, reducing the test run
time by 50%-60%.