ActorTaskScheduler QueueTask for LongRunning #1410
Its not a bug, its a missaligment to the normal behavior of .net. If
something is creating a task with the TaskCreationOptions.LongRunning
Flag it assumes that it can hold the thread for a long running process.
If this process is a never ending message loop, it will lock one thread
and cell/actor in akka.