Fixes handling of messges intended for another node
Given the following setup:
- Node: client (cluster client using remoting)
- Node: lighthouse (seed node for cluster)
- Node: worker (cluster member)

where both `lighthouse` and `worker` start `ClusterClientReceptionist`
but only the `worker` node registers a service (`/user/worker`).

If the `client` node uses the `lighthouse` node as the only value in
`initial-contacts` and attempts to send a message to `/user/worker`, the
log on the `lighthouse` node indicates that it received the message, but
the `worker` node never does.

With this change, the `worker` node will now receive the message.