Fix Serialization KeyNotFoundException when looking up serializers (#2356)
* Fix Serialization KeyNotFoundException when looking up serializers.

A KeyNotFoundException can occur if the serializer dll is not present
in the search paths for binaries. This can occure when there is a
proj A that depends on B and B depends on Akka. There is no direct
reference to any of the serializers in either A or B besides being
referenced in HOCON. MSBuild likes to optimize dlls out of the out-
put if they are not directly referenced in code.

* Better use of .TryGetValue()