Messages are logged recursively with custom logger. (#2379)
* Messages are logged recursively with custom logger.

When "loglevel = DEBUG" and "receive = on" are set in HOCON and custom
logger is in use, messages are logged recursively. In addition, some of
the messages are not implementing IAutoReceivedMessage, although seem to
be sent to Self, causing the log to be noisy.

* Messages are logged recursively with custom logger.

Rolled back implementations of IAutoReceivedMessage for
AckIdleCheckTimer, Prune, HeartbeatTick, ReapUnreachableTick and
INoSerializationVerificationNeeded classes (broke specs). Slight
modification to ActorCell.DefaultMessages.ReceiveMessage.

* Messages are logged recursively with custom logger.

Using string concatenation instead of string.Format (attempt to improve
performance)

* Mssages are logged recursively with custom logger

Reversed the changes, as this is not a bug in ActorCell.

* Messages are logged recursively with custom logger

Refactored output to exclude the receiver part (can be logged via
LogEvent)