Updating ProtobufSerializer to automatically lookup the MessageParser (#3032)
* Updating ProtobufSerializer to automatically lookup parser

The existing implementation of ProtobufSerializer requires RegisterFileDescriptor to be called with the file descriptor of every protobuf message that will be serialized.  This adds a significant amount of boilerplate to serializer protobuf messages automatically.  To work around this, I've updated the fromBinary to lookup the MessageParser at runtime based on the type.  This brings it inline with how the JVM Akka protobuf serializer works.

* Fixing approved remote api