[xml-docs] Fixed broken xrefs / missing field refs
This PR fixes a number of broken cross references and adds a couple of missing parameter references.

Most of the cross references were referencing fields in the subclass. For some reason MSBuild doesnt like that. The fix was to reference the fields of the base class and the link text pointed to the correct subclass field.