[build] Fix CS0162 warnings
This PR fixes a number of warnings related to CS0162 (Unreachable code detected).

Most of these are related to conditional debugging flags that aren't necessarrily set during a normal compilation. There are some however that are there as to clear any compile errors that would happen due to ambigious references.

The pragmas are added to the function boundaries so as not to pollute the logic in the function.