Update tutorial-1.md
Added a standalone, working code example for the `Structure of an IActorRef and Paths of Actors` section. Looking at the documentation trend, it should ideally be made a into a reference to a github-hosted piece of code. Looking at how tutorials are organised at the moment, the structure doesn't really encourage individual contributions - the linked examples are a part of a larger project and a larger group of examples, which are not standalone. To contribute, one would have to build the entire solution and modify the entire `Akka.Docs.Tutorials.csproj` project. What I would advise would encourage community contributions/engagement are more standalone, solution-independent examples that illustrate sections/problems clearer.