Akka.Streams: memory-optimize `ActorMaterializer` HOCON injection (#6440)
* optimize `ActorMaterializer` to only inject HOCON configuration once

creating lots of materializer instances (i.e. in actors that own their own streams) results in a tremendous amount of unreclaimable HOCON-related memory - on the order of 11Gb in https://github.com/Aaronontheweb/AkkaSqlQueryCrushTest

This tries to avoid injecting the top level Akka.Streams fallback every time a materializer is used. In addition to that, we try to cache the default materializer settings parsed from the `ActorSystem`s HOCON if none are provided.

* modernized syntax

* improve cache to detect multiple `ActorSystem` instances